Get shooter to stop and add camera servo
Removed the CIM Motor 2 and associated references (we'll just use one
PWM port and swap the wires for one).
Also removed joystick 1 shoot the ball button as it was redundant. added
the camera system with a servo along with swapview command.
changed some requires in subsystems. Uncommented the stop the shooter
commands.