Sybsystem changes
Merged shooting subsystem and pic up sybsystem together sicne they are
going to be used in the same command. Pickup command takes the ball in
and then stops, and then the shooter command will turn the pick up
command as well as the feeder and the shooter to shoot the ball.